[DF-10924] Build fix
Former-commit-id: 54bfe7ac9818263eced36118d00f4942228bd04b

+++ Sources/Internal/Platform/TemplateWin32/PrivateTextFieldWinUAP.cpp
@@ -729,7 +729,7 @@ void PrivateTextFieldWinUAP::OnTextChanged()
         // If delegate changes text, set new caret position on the end
         if (curText != oldText)
         {
-            newCaretPos = curText.length();
+            newCaretPos = static_cast<int32>(curText.length());
         }
     });
